% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_gd_optim.R
\name{linear_gd_optim}
\alias{linear_gd_optim}
\title{linear_gd_optim}
\usage{
linear_gd_optim(
  par,
  X,
  Y,
  tolerance = 0.001,
  maxit = 1000,
  stepsize = 0.001,
  verbose = T
)
}
\arguments{
\item{par}{Initial values.}

\item{X}{Data predictiors.}

\item{Y}{Response variables.}

\item{tolerance}{Value for the stopping criterion. The default value is set to 1e-3.}

\item{maxit}{Maximum iterations allowed. The default value is set to 1000.}

\item{stepsize}{Length of the stepsize parameter. The default value is set to 1e-3.}

\item{verbose}{If set TRUE the function produce messages during the computation.}
}
\value{
Vector containing the estimated parameters.
}
\description{
This function provides a personalized way for optimizing functions
}
