% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_grad.R
\name{my_grad}
\alias{my_grad}
\title{Gradient computation}
\usage{
my_grad(par, X, Y, verbose = T)
}
\arguments{
\item{par}{Parameters values.}

\item{X}{Data predictiors.}

\item{Y}{Response variables.}
}
\value{
Returns the vector containing the gradient scores.
}
\description{
This function provides a vectorized version of the gradient computation
}
\examples{

# Having the parameters vector "par", the predictors matrix "X",
# and the outcome vector "Y", the standard usage for computing the
# descend gradient direction is the following:

my_grad(par, X, Y)

}
