% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_grad_old.R
\name{my_grad_old}
\alias{my_grad_old}
\title{Gradient computation}
\usage{
my_grad_old(par, X, Y, verbose = T)
}
\arguments{
\item{par}{Parameters values.}

\item{X}{Data predictiors.}

\item{Y}{Response variables.}

\item{verbose}{If set TRUE the function produce messages during the computation.}
}
\value{
Vector containing the gradient scores.
}
\description{
This function provides a basic way to compute gradient
}
