% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_hess.R
\name{my_hess}
\alias{my_hess}
\title{Hessian computation}
\usage{
my_hess(X, verbose = T)
}
\arguments{
\item{X}{Data predictors.}
}
\value{
Vector containing the Hessian scores.
}
\description{
This function provides a vectorized version of the Hessian computation
}
