% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_k_fold_cv.R
\name{my_k_fold_cv}
\alias{my_k_fold_cv}
\title{K-fold cross validation for linear model}
\usage{
my_k_fold_cv(par, X, Y, K, method = "gd", parallel = FALSE)
}
\arguments{
\item{par}{Initial values.}

\item{X}{Data matrix predictors.}

\item{Y}{Response vector.}

\item{K}{Number of folds.}

\item{method}{Method: "gd" for gradient descent, "sd" for steepest descent.}

\item{parallel}{Apply the parallel version on not: TRUE for parallel version, FALSE for sequential version.}
}
\value{
Returns the k-fold cross validation error.
#'
}
\description{
K-fold cross validation for linear model
}
\examples{
library(myOpt)
## basic example code

set.seed(8675309)

# data simulation for example purposes

n = 1000

x1 = rnorm(n)
x2 = rnorm(n)
X <- cbind(rep(1,n),x1,x2)

Y = 1 + 0.5*x1 + 0.2*x2 + rnorm(n)

# random initial values for the parameters of the linear model

par <- rnorm(dim(X)[2])

# number of folds

K <- 5

# the function returns the k-fold cross validation error

err_kfold <- my_k_fold_cv(par, X, Y, K, method="sd", parallel=TRUE)

}
