% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_linear_gd_predict.R
\name{my_linear_gd_predict}
\alias{my_linear_gd_predict}
\title{Prediction of a linear model}
\usage{
my_linear_gd_predict(par, X)
}
\arguments{
\item{par}{Parameters vector to be used in the prediction.}

\item{X}{Data matrix predictiors.}
}
\value{
Returns the vector containing the predictions computed for the data contained in X
}
\description{
Prediction of a linear model
}
\examples{


}
