% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_linear_predict.R
\name{my_linear_predict}
\alias{my_linear_predict}
\title{Title}
\usage{
my_linear_predict(par, X)
}
\arguments{
\item{par}{Parameters vector to be used in the prediction.}

\item{X}{Data matrix predictiors.}
}
\value{
Vector containing the predictions computed for the data contained in X
}
\description{
Title
}
\examples{


}
