% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pred_error.R
\name{my_pred_error}
\alias{my_pred_error}
\title{Title}
\usage{
my_pred_error(Y, Y_pred)
}
\arguments{
\item{Y}{Response vector.}

\item{Y_pred}{Prediction of the response vector.}
}
\value{
Returns the mean squared error of the prediction.
}
\description{
Title
}
