% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_k_fold_cv_gd.R
\name{seq_k_fold_cv_gd}
\alias{seq_k_fold_cv_gd}
\title{K-fold cross validation for linear model}
\usage{
seq_k_fold_cv_gd(par, X, Y, K)
}
\arguments{
\item{par}{Initial values.}

\item{X}{Data matrix predictiors.}

\item{Y}{Response vector.}

\item{K}{Number of folds.}
}
\value{
Returns the k-fold cross validation error.
}
\description{
K-fold cross validation for linear model
}
