% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_gd_optim.R
\name{linear_gd_optim}
\alias{linear_gd_optim}
\title{linear_gd_optim}
\usage{
linear_gd_optim(
  par,
  X,
  Y,
  tolerance = 1e-06,
  maxit = 10000,
  stepsize = 0.001,
  verbose = T
)
}
\arguments{
\item{par}{Initial values.}

\item{X}{Data matrix predictiors.}

\item{Y}{Response variables.}

\item{tolerance}{Value for the stopping criterion. The default value is set to 1e-6.}

\item{maxit}{Maximum iterations allowed. The default value is set to 10000.}

\item{stepsize}{Length of the stepsize parameter. The default value is set to 1e-3.}

\item{verbose}{If set TRUE the function produce messages during the computation.}
}
\value{
Vector containing the estimated parameters.
}
\description{
This function provides a personalized efficient way for optimizing functions
}
\examples{
library(myOpt)
## basic example code

set.seed(8675309)

# data simulation for example purposes

n = 1000

x1 = rnorm(n)
x2 = rnorm(n)
X <- cbind(rep(1,n),x1,x2)

Y = 1 + 0.5*x1 + 0.2*x2 + rnorm(n)

# random initial values for the parameters of the linear model

par <- rnorm(3)


est_par <- linear_gd_optim(par, X, Y)

}
