% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_sd_optim.R
\name{linear_sd_optim}
\alias{linear_sd_optim}
\title{linear_sd_optim}
\usage{
linear_sd_optim(par, X, Y, tolerance = 1e-06, maxit = 10000, verbose = T)
}
\arguments{
\item{par}{Initial values.}

\item{X}{Data predictors.}

\item{Y}{Response variables.}

\item{tolerance}{Value for the stopping criterion. The default value is set to 1e-6.}

\item{maxit}{Maximum iterations allowed. The default value is set to 10000.}

\item{verbose}{If set TRUE the function produce messages during the computation.}
}
\value{
Vector containing the estimated parameters.
}
\description{
This function provides a personalized efficient way for optimizing functions
}
