% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_grad.R
\name{my_grad}
\alias{my_grad}
\title{linear_gd_optim}
\usage{
my_grad(par, X, Y)
}
\arguments{
\item{par}{Parameters values.}

\item{X}{Data predictiors.}

\item{Y}{Response variables.}
}
\value{
Vector containing the gradient scores.
}
\description{
This function provides a personalized way for optimizing functions
}
