% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_grad.R
\name{my_grad}
\alias{my_grad}
\title{Gradient computation}
\usage{
my_grad(par, X, Y, verbose = T)
}
\arguments{
\item{par}{Parameters values.}

\item{X}{Data predictors.}

\item{Y}{Response variables.}

\item{verbose}{If set TRUE the function produce messages during the computation.}
}
\value{
Returns the vector containing the gradient scores.
}
\description{
This function provides a vectorized version of the gradient computation
}
\examples{

# Having the parameters vector "par", the predictors matrix "X",
# and the outcome vector "Y",
# the gradient is calculated as follows:

set.seed(8675309)

# data simulation for example purposes

n = 1000

x1 = rnorm(n)
x2 = rnorm(n)
X <- cbind(rep(1,n),x1,x2)

Y = 1 + 0.5*x1 + 0.2*x2 + rnorm(n)

# random initial values for the parameters of the linear model

par <- rnorm(dim(X)[2])

# gradient calculation

my_grad(par, X, Y)

}
